Added lives to the player.
Added the extra lives to the HUD

Added regions around the individual animation sets. The remaining animations sets for enemies and ammunition still need to be defined.

Implemented 1Ups (Extra Live events) as pickups. They add 2000 points to the score, and 1 life to the player.

Added Animation defines for Springs

Added logic to retain the player's lives when restarting the level

Added logic to decrement player lives on death, and close the game when the player runs out of lives.