Added a setting for showing the framerate on the screen
Fixed some issues with horizontal spring behavior.

Added some optimizations with shaders. I replaced all calls to glUseProgram with calls to useProgram. This new method uses a cached copy of the last value passed to glUseProgram to determine if this call to glUseProgram needs to be executed.

Added setUniform methods as proxies to the glUniform methods. These use the currentShader and gets the uniform location from a cache so that I don't need to call glGetUniformLocation all the time.

Added a method to get and cache the uniform locations for the shaders.
Added calls to this method for each shader we use.
Changed the way that the stringJoin function was implemented. Hopefully using a stringstream will be more efficient.