Added a null check to the Actor class when getting the AnimationFrame for checking collision. If there's no animation frame, there cannot be a collision.
Changed the depths in the VBOs to be layer based, with the depth being 8 - layer, and that makes the sprite layer 5.

Got the depth buffer working in the FBO, so now I can render the sprite layer after I render the level. This simplifies the rendering code.

I separated the rending and updating logic into two functions separate from the run loop. This will make further changes to the game logic a little easier.

Fixed the framerate issue when the game is not running with VSYNC on.
Enabled VSYNC by default.